Done Mathe Aufagabe class with the main functinalities

+++ MatheAufgabe.java
@@ -5,7 +5,7 @@ public class MatheAufgabe {
     public double sub(double x, double y) {
         return x - y;
     }
-    public double mul(double x, double y) {
+    public int mul(double x, double y) {
         int result=0;
         for(int i=0; i<=y; i++) {
             result+=x;
